Climate models upload

Automatically load climate models (netCDF/NcML) in a tidy format.

load_data(
  path.to.rcps,
  country,
  variable,
  xlim = NULL,
  ylim = NULL,
  years.proj,
  years.hist,
  path.to.obs = NULL,
  n.cores = NULL,
  buffer = 0,
  domain = NULL
)

Arguments

path.to.rcps
Path to the directory containing the RCPs/SSPs folders and historical simulations. For example, home/user/data/. data would contain subfolders with the climate models. Historical simulations have to be contained in a folder called historical. If path.to.rcps is set as CORDEX-CORE, CORDEX-CORE simulations from RCM RegCM4 will be loaded

country
A character string, in english, indicating the country of interest. To select a bounding box, set country to NULL and define arguments xlim and ylim

variable
A character string indicating the variable

xlim
Vector of length = 2, with minimum and maximum longitude coordinates, in decimal degrees, of the bounding box selected.

ylim
Same as xlim, but for the selection of the latitudinal range.

years.proj
Numerical range, years to select for projections

years.hist
Numerical range, years to select for historical simulations and observations

path.to.obs
Default to NULL, if not, indicate the absolute path to the directory containing a reanalysis dataset, for example ERA5. To automatically load W5E5. specify W5E5

n.cores
Integer, number of cores to use in parallel processing, default is 9

buffer
Numeric. Default is zero.

domain
Specify the CORDEX-CORE domain (e.g AFR-22, EAS-22). Used with path.to.rcps = CORDEX-CORE. Default is NULL

Value

Tibble with column list

Examples

fpath <- system.file("extdata/", package="cavaR")
exmp1 <- load_data(country = "Moldova", variable="hurs", years.hist=2000, years.proj=2010,
              path.to.rcps = fpath)

#> Your directory contains the following folders: 
#> /tmp/RtmplkDjoP/temp_libpath2e7097f8699b7/cavaR/extdata//historical 
#> /tmp/RtmplkDjoP/temp_libpath2e7097f8699b7/cavaR/extdata//rcp2.6 
#> /tmp/RtmplkDjoP/temp_libpath2e7097f8699b7/cavaR/extdata//rcp8.5 
#> all files within the listed folders will be uploaded 

#> Warning: if you do not specify a reanalysis/observational gridded dataset, bias-correction cannot be performed. To load W5E5 set path.to.obs as W5E5 

#> The process is currently parallelized using 9 cores. Type TRUE to continue or set the argument n.cores 
#> 

#> 2022-11-20 19:06:49 Loading /tmp/RtmplkDjoP/temp_libpath2e7097f8699b7/cavaR/extdata//historical/EUR-22_MOHC-HadGEM2-ES_historical_r1i1p1_GERICS-REMO2015_v1_day_1970_2005_cr.nc

#> 2022-11-20 19:06:49 Done

#> 2022-11-20 19:06:51 Loading /tmp/RtmplkDjoP/temp_libpath2e7097f8699b7/cavaR/extdata//historical/EUR-22_MPI-M-MPI-ESM-LR_historical_r1i1p1_GERICS-REMO2015_v1_day_1971-2005_cr.nc

#> 2022-11-20 19:06:51 Done

#> 2022-11-20 19:06:53 Loading /tmp/RtmplkDjoP/temp_libpath2e7097f8699b7/cavaR/extdata//historical/EUR-22_NCC-NorESM1-M_historical_r1i1p1_GERICS-REMO2015_v1_day_1970_2005_cr.nc

#> 2022-11-20 19:06:53 Done

#> 2022-11-20 19:06:51 Loading /tmp/RtmplkDjoP/temp_libpath2e7097f8699b7/cavaR/extdata//rcp2.6/EUR-22_MOHC-HadGEM2-ES_rcp26_r1i1p1_GERICS-REMO2015_v1_day_2006_2100_cr.nc

#> 2022-11-20 19:06:51 Done

#> 2022-11-20 19:06:53 Loading /tmp/RtmplkDjoP/temp_libpath2e7097f8699b7/cavaR/extdata//rcp2.6/EUR-22_MPI-M-MPI-ESM-LR_rcp26_r1i1p1_GERICS-REMO2015_v1_day_2006_2100_cr.nc

#> 2022-11-20 19:06:53 Done

#> 2022-11-20 19:06:55 Loading /tmp/RtmplkDjoP/temp_libpath2e7097f8699b7/cavaR/extdata//rcp2.6/EUR-22_NCC-NorESM1-M_rcp26_r1i1p1_GERICS-REMO2015_v1_day_2006_2100_cr.nc

#> 2022-11-20 19:06:55 Done

#> 2022-11-20 19:06:53 Loading /tmp/RtmplkDjoP/temp_libpath2e7097f8699b7/cavaR/extdata//rcp8.5/EUR-22_MOHC-HadGEM2-ES_rcp85_r1i1p1_GERICS-REMO2015_v1_day_2006_2100_cr.nc

#> 2022-11-20 19:06:53 Done

#> 2022-11-20 19:06:55 Loading /tmp/RtmplkDjoP/temp_libpath2e7097f8699b7/cavaR/extdata//rcp8.5/EUR-22_MPI-M-MPI-ESM-LR_rcp85_r1i1p1_GERICS-REMO2015_v1_day_2006_2100_cr.nc

#> 2022-11-20 19:06:55 Done

#> 2022-11-20 19:06:56 Loading /tmp/RtmplkDjoP/temp_libpath2e7097f8699b7/cavaR/extdata//rcp8.5/EUR-22_NCC-NorESM1-M_rcp85_r1i1p1_GERICS-REMO2015_v1_day_2006_2100_cr.nc

#> 2022-11-20 19:06:57 Done

#> 
#>  2022-11-20 19:06:57 Aggregating members 

#> 2022-11-20 19:06:57 Done

exmp2 <- load_data(country = "Somalia", variable="tas", years.hist=2000, years.proj=2010,
              path.to.rcps = "CORDEX-CORE", domain="AFR-22")

#> Warning: if you do not specify a reanalysis/observational gridded dataset, bias-correction cannot be performed. To load W5E5 set path.to.obs as W5E5 

#> The process is currently parallelized using 9 cores. Type TRUE to continue or set the argument n.cores 
#> 

#> 2022-11-20 19:07:02 Loading CORDEX-AFR-22_MOHC-HadGEM2-ES_historical_r1i1p1_ICTP-RegCM4-7_v0

#> 2022-11-20 19:07:16 Done

#> 2022-11-20 19:07:04 Loading CORDEX-AFR-22_MPI-M-MPI-ESM-MR_historical_r1i1p1_ICTP-RegCM4-7_v0

#> 2022-11-20 19:07:18 Done

#> 2022-11-20 19:07:06 Loading CORDEX-AFR-22_NCC-NorESM1-M_historical_r1i1p1_ICTP-RegCM4-7_v0

#> 2022-11-20 19:07:21 Done

#> 2022-11-20 19:07:04 Loading CORDEX-AFR-22_MOHC-HadGEM2-ES_rcp26_r1i1p1_ICTP-RegCM4-7_v0

#> 2022-11-20 19:07:23 Done

#> 2022-11-20 19:07:06 Loading CORDEX-AFR-22_MPI-M-MPI-ESM-MR_rcp26_r1i1p1_ICTP-RegCM4-7_v0

#> 2022-11-20 19:07:26 Done

#> 2022-11-20 19:07:08 Loading CORDEX-AFR-22_NCC-NorESM1-M_rcp26_r1i1p1_ICTP-RegCM4-7_v0

#> 2022-11-20 19:07:33 Done

#> 2022-11-20 19:07:06 Loading CORDEX-AFR-22_MOHC-HadGEM2-ES_rcp85_r1i1p1_ICTP-RegCM4-7_v0

#> 2022-11-20 19:07:26 Done

#> 2022-11-20 19:07:08 Loading CORDEX-AFR-22_MPI-M-MPI-ESM-MR_rcp85_r1i1p1_ICTP-RegCM4-7_v0

#> 2022-11-20 19:07:28 Done

#> 2022-11-20 19:07:09 Loading CORDEX-AFR-22_NCC-NorESM1-M_rcp85_r1i1p1_ICTP-RegCM4-7_v0

#> 2022-11-20 19:07:30 Done

#> 
#>  2022-11-20 19:07:33 Aggregating members 

#> 2022-11-20 19:07:34 Done